+++ src/CustomPropertyTypes/register.ts
@@ -29,7 +29,7 @@ import { multiSelectPropertyType } from "./MultiSelect";
 import { numericPropertyType } from "./Numeric";
 import { arrayPropertyType } from "./Array";
 import { Icon } from "~/lib/types/icons";
-import { relationPropertyType } from "./Relation";
+// import { relationPropertyType } from "./Relation";
 
 export const getCustomPropertyTypesArr = (): CustomPropertyType[] => [
 	arrayPropertyType,
@@ -42,7 +42,7 @@ export const getCustomPropertyTypesArr = (): CustomPropertyType[] => [
 	objectPropertyType,
 	colorPropertyType,
 	ratingPropertyType,
-	relationPropertyType,
+	// relationPropertyType,
 	dateCustomPropertyType,
 	sliderPropertyType,
 	timePropertyType,