+++ src/Plugin/plugin.ts
@@ -1,14 +1,4 @@
-import {
-	Modal,
-	ButtonComponent,
-	Plugin,
-	MarkdownPostProcessorContext,
-	Component,
-	MarkdownRenderChild,
-	MarkdownRenderer,
-	App,
-	MarkdownView,
-} from "obsidian";
+import { Modal, ButtonComponent, Plugin, MarkdownView } from "obsidian";
 import {
 	BetterPropertiesSettings,
 	betterPropertiesSettingsSchema,
@@ -51,13 +41,6 @@ export class BetterProperties extends Plugin {
 			this.rebuildLeaves();
 		});
 		this.handlePropertyLabelWidth();
-
-		this.registerMarkdownCodeBlockProcessor("script", (source, el, ctx) => {
-			new Script(this, el, source, ctx);
-		});
-		window.CodeMirror.defineMode("script", (config) =>
-			window.CodeMirror.getMode(config, "javascript")
-		);
 
 		registerBpJsCodeProcessors(this);
 		this.registerEvent(
@@ -215,97 +198,5 @@ export class BetterProperties extends Plugin {
 		const newSettings = cb(this.settings);
 		this.settings = { ...newSettings };
 		await this.saveSettings();
-	}
-}
-
-// TODO move to a separate plugin and make inline codeblock renderer
-class Script {
-	public app: App;
-	public mdrc: MarkdownRenderChild;
-	public component: Component;
-	public containerEl: HTMLElement;
-	constructor(
-		public plugin: BetterProperties,
-		public el: HTMLElement,
-		public source: string,
-		public ctx: MarkdownPostProcessorContext
-	) {
-		this.app = plugin.app;
-		this.mdrc = new MarkdownRenderChild(el);
-		this.component = new Component();
-		this.containerEl = el.createDiv();
-		this.mdrc.addChild(this.component);
-		this.ctx.addChild(this.mdrc);
-
-		if (!source) {
-			this.renderHelp();
-			return;
-		}
-		this.runCode(source);
-	}
-
-	async renderHelp() {
-		const helpText = `## Usage\n - Enter \`console.log(script)\` to see what information is available to use in your script\n\n- To load external scripts: \`script.loadScript("path/to/main.js", "path/to/styles.css")\`\n\n## Typescript\n\`\`\`\ndeclare const script: {\n  app: App;\n  mdrc: MarkdownRenderChild;\n  component: Component;\n  plugin: Plugin;\n  el: HTMLElement;\n  source: string;\n  ctx: MarkdownPostProcessorContext;\n};\n\`\`\``;
-		const container = this.el.createDiv();
-
-		await MarkdownRenderer.render(
-			this.plugin.app,
-			helpText,
-			container,
-			this.ctx.sourcePath,
-			this.component
-		);
-	}
-
-	async runCode(code: string) {
-		try {
-			const func = eval(`async (script) => {${code}}`);
-			this.containerEl.empty();
-			await func(this);
-		} catch (e) {
-			const msg = e instanceof Error ? e.message : "Unknown Error";
-			const fullMsg =
-				"An error occurred when rendering script codeblock: " + msg;
-			console.error(fullMsg);
-			this.el.textContent = fullMsg;
-		}
-	}
-
-	async loadScript(pathToMainJs: string, pathToStylesCss?: string) {
-		const {
-			plugin: {
-				app: { vault },
-			},
-			el,
-		} = this;
-		const mainJsFile = vault.getFileByPath(pathToMainJs);
-		if (!mainJsFile) {
-			throw new Error("main.js file not found by path: " + pathToMainJs);
-		}
-
-		const stylesCssFile = pathToStylesCss
-			? vault.getFileByPath(pathToStylesCss)
-			: null;
-		if (pathToStylesCss && !stylesCssFile) {
-			throw new Error("styles.css file not found by path: " + pathToStylesCss);
-		}
-
-		if (stylesCssFile) {
-			const styleEl = el.createEl("style");
-			styleEl.innerHTML = await vault.read(stylesCssFile);
-		}
-
-		await this.runCode(await vault.read(mainJsFile));
-	}
-
-	obsidian() {
-		return require("obsidian");
-	}
-
-	async refresh() {
-		this.component.unload();
-		this.el.empty();
-		this.containerEl = this.el.createDiv();
-		await this.runCode(this.source);
 	}
 }